feat: support streamed execution with cdf
